rename to `div_norm`
