+++ src/prfmodel/models/__init__.py
@@ -9,10 +9,10 @@ from .compressive_spatial_summation import Gaussian2DCSSPRFModel
 from .compressive_spatial_summation import init_css_from_gaussian
 from .difference_of_gaussians import DoG2DPRFModel
 from .difference_of_gaussians import init_dog_from_gaussian
-from .divisive_normalization import DivNormGaussian2DPRFModel
-from .divisive_normalization import DivNormPRFModel
-from .divisive_normalization import init_dn_from_dog
-from .divisive_normalization import init_dn_from_gaussian
+from .div_norm import DivNormGaussian2DPRFModel
+from .div_norm import DivNormPRFModel
+from .div_norm import init_dn_from_dog
+from .div_norm import init_dn_from_gaussian
 from .encoding import CFStimulusEncoder
 from .encoding import CompressiveEncoder
 from .encoding import PRFStimulusEncoder
